requirements.txt
python_dotenv
aiohttp
pandas
